+++ src/pages/Home.tsx
@@ -1,11 +1,11 @@
 import seedrandom from 'seedrandom'
 import styled from 'styled-components/macro'
-import Button from '../common/Button'
+import Groups from '../common/Groups'
 import Rooms from '../common/Rooms'
-import ShareableLink from '../common/ShareableLink'
+import Seed from '../common/Seed'
 import { distribute, shuffle } from '../services/list'
 
-interface HomeProps {
+interface Props {
   rooms: string[]
   students: string[]
   seed: string
@@ -21,58 +21,23 @@ export default function Home({
   onSeedChange,
   groups,
   onGroupsChange,
-}: HomeProps) {
+}: Props) {
   const rng = seedrandom(seed)
   const studentGroups = distribute(shuffle(seed, students), groups)
 
   return (
     <Main>
-      <section>
-        <h2>Seed</h2>
-        <Input
-          type="text"
-          value={seed}
-          onChange={({ target }) => onSeedChange(target.value)}
-        />
-        <Button onClick={() => onSeedChange(new Date().toLocaleDateString())}>
-          daily
-        </Button>
-        <Button onClick={() => onSeedChange(rng().toString().slice(2))}>
-          random
-        </Button>
-      </section>
-      <section>
-        <h2>Number of Groups</h2>
-        <Button
-          disabled={groups === 1}
-          onClick={() => onGroupsChange(groups - 1)}
-        >
-          less
-        </Button>
-        <span className="groups"> {groups} </span>
-        <Button
-          disabled={groups === rooms.length}
-          onClick={() => onGroupsChange(groups + 1)}
-        >
-          more
-        </Button>
-      </section>
-      <section>
-        <h2>Rooms</h2>
-        <ShareableLink rooms={rooms} groups={studentGroups} />
-        <Rooms rooms={rooms} groups={studentGroups} />
-      </section>
+      <Seed seed={seed} onChange={onSeedChange} rng={rng} />
+      <Groups
+        groups={groups}
+        maxGroups={rooms.length}
+        onChange={onGroupsChange}
+      />
+      <Rooms rooms={rooms} studentGroups={studentGroups} />
     </Main>
   )
 }
 
-const Input = styled.input`
-  padding: 8px;
-  font-size: 1em;
-  font-family: inherit;
-  border: 1px solid white;
-  color: var(--color-blue);
-`
 const Main = styled.main`
   display: grid;
   grid-template-columns: 1fr 1fr;
